refactor: moved fetch note unit test to `tests/note`
